[dt,pwrmgr] Move pwrmgr dt binding to hw/top/dt instead of ipgening it
