+++ hw/top/dt/pwrmgr_binding.py
@@ -12,7 +12,7 @@ import os
 import sys
 
 sys.path.insert(0, os.path.dirname(os.path.realpath(__file__)))
-from ipconfig import PwrmgrIpConfig  # noqa: E402
+from pwrmgr_ipconfig import PwrmgrIpConfig  # noqa: E402
 
 HEADER_EXT_TEMPLATE = """
 /**
